+++ src/args.rs
@@ -88,7 +88,9 @@ impl<'a> Args<'a> {
                 "--version" => res.set_action(Action::Version)?,
                 "-c" | "--create" => res.set_action(Action::Create)?,
                 "-a" | "--alias" => res.set_alias(args.next_arg()?)?,
-                "-C" | "--config" | "--configure" => res.set_action(Action::Config)?,
+                "-C" | "--config" | "--configure" => {
+                    res.set_action(Action::Config)?
+                }
                 "-t" | "--template" => res.set_template(args.next_arg()?)?,
                 "-d" | "--directory" => res.set_path(args.next_arg()?)?,
                 "-cf" | "--create-from" => res.set_action_template_path(